requirements.txt
pygame==2.5.2
moviepy
Pillow==9.4.0
tkinter
socket
json